requirements.txt
Django>=3.0,<4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1

# flake8>=3.6.0,<3.7.0